day148/nlp_query_system/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
spacy==3.7.4
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
redis==5.0.4
psycopg2-binary==2.9.9
python-dateutil==2.9.0
pydantic==2.7.1
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
python-multipart==0.0.9
aiofiles==23.2.1
